06_train/extras/bert/fast-bert/requirements.txt
pytorch-lamb
tensorboardX
fastprogress
sklearn
spacy
transformers>=2.3.0
pandas
python-box
tokenizers